mmr-testing/mediasource-endofstream-forMMR.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Calls to MediaSource.endOfStream() without error</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="mediasource-util.js"></script>
<script>
    mediasource_test(function(test, mediaElement, mediaSource)
    {
        mediaSource.duration = 2;
        mediaSource.endOfStream();
        //hzhong
        //assert_equals(mediaSource.duration, 0);
        test.done();
    }, 'MediaSource.endOfStream(): duration truncated to 0 when there are no buffered coded frames');

</script>
